requirements.txt
python-dotenv~=1.0.0
pyTelegramBotAPI~=4.12.0
cloudwatch~=1.0.5
